2.04

+++ inc/delete_dir.php
@@ -3,9 +3,15 @@
   // Удаление непустых директорий:
 function deleteDir($dir)
 {
-  return is_file($dir)?
-  @unlink($dir):
-  array_map('deleteDir',glob($dir.'/*'))==@rmdir($dir);
-  // echo  'Удалить '.$$dir.'?';
+   if ($dir != $_SERVER['DOCUMENT_ROOT'])
+      {
+        return is_file($dir)?
+        @unlink($dir):
+        array_map('deleteDir',glob($dir.'/*'))==@rmdir($dir);
+      }
+   else
+      {
+         echo  'Неверный каталог '.$dir;
+         return false;
+      }
 }
- ?>
